CMParts: Properly fix auto-power-save string format exception
We were originally encountering an issue when formatting
the auto-power-save ListPreference summary.

As detailed by Danny Baumann, when having the summary string
formatted with '%d %% foo bar', updateAutoPowerSaveSummary()
would format it to '10 % foo bar'. When setting it to the preference,
the getter would run '10 % foo bar' through String.format()
again. This would result in the issue detailed in NIGHTLIES-3540.

To properly fix this issue, have the summary string formatted
with '%s foo bar', where %s is '%d' + '%' in this case.

Change-Id: Ia90a75f40fddb6e7ab6a01603d40fda7263cba4c
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>